Make build time better, add option for --check
